Uniformly formatted solidity sources

+++ celo/audited-sources/celo-org/celo-monorepo/packages/protocol/contracts-0.8/common/interfaces/IWETH.sol/89fa8f2660c0dde368cbbceddb0f25b77ffd7e1c/IWETH.sol
@@ -2,16 +2,16 @@
 pragma solidity >=0.8.7 <0.8.20;
 
 interface IWETH {
-  function deposit() external payable;
-  function withdraw(uint wad) external;
+    function deposit() external payable;
+    function withdraw(uint256 wad) external;
 
-  function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
 
-  function approve(address guy, uint wad) external returns (bool);
+    function approve(address guy, uint256 wad) external returns (bool);
 
-  function transfer(address dst, uint wad) external returns (bool);
+    function transfer(address dst, uint256 wad) external returns (bool);
 
-  function transferFrom(address src, address dst, uint wad) external returns (bool);
+    function transferFrom(address src, address dst, uint256 wad) external returns (bool);
 
-  function allowance(address owner, address spender) external returns (uint256);
+    function allowance(address owner, address spender) external returns (uint256);
 }
