+++ roninnetwork/audited-sources/axieinfinity/ronin-dpos-contracts/contracts/1d3f5e3c1de471edd6e8b4ea15167130f40e3d90/mocks/validator/MockRoninValidatorSetOverridePrecompile.sol
@@ -2,46 +2,42 @@
 
 pragma solidity ^0.8.9;
 
-import "../MockPrecompile.sol";
 import "../../ronin/validator/RoninValidatorSet.sol";
+import "../MockPrecompile.sol";
 
 contract MockRoninValidatorSetOverridePrecompile is RoninValidatorSet, MockPrecompile {
-  constructor() {}
+    constructor() {}
 
-  function arrangeValidatorCandidates(
-    address[] memory _candidates,
-    uint256[] memory _trustedWeights,
-    uint _newValidatorCount,
-    uint _maxPrioritizedValidatorNumber
-  ) external pure returns (address[] memory) {
-    _arrangeValidatorCandidates(_candidates, _trustedWeights, _newValidatorCount, _maxPrioritizedValidatorNumber);
-    return _candidates;
-  }
+    function arrangeValidatorCandidates(
+        address[] memory _candidates,
+        uint256[] memory _trustedWeights,
+        uint256 _newValidatorCount,
+        uint256 _maxPrioritizedValidatorNumber
+    ) external pure returns (address[] memory) {
+        _arrangeValidatorCandidates(_candidates, _trustedWeights, _newValidatorCount, _maxPrioritizedValidatorNumber);
+        return _candidates;
+    }
 
-  function _pcSortCandidates(address[] memory _candidates, uint256[] memory _weights)
-    internal
-    pure
-    override
-    returns (address[] memory _result)
-  {
-    return sortValidators(_candidates, _weights);
-  }
+    function _pcSortCandidates(address[] memory _candidates, uint256[] memory _weights)
+        internal
+        pure
+        override
+        returns (address[] memory _result)
+    {
+        return sortValidators(_candidates, _weights);
+    }
 
-  function _pcPickValidatorSet(
-    address[] memory _candidates,
-    uint256[] memory _weights,
-    uint256[] memory _trustedWeights,
-    uint256 _maxValidatorNumber,
-    uint256 _maxPrioritizedValidatorNumber
-  ) internal pure override returns (address[] memory _result, uint256 _newValidatorCount) {
-    _result = pickValidatorSet(
-      _candidates,
-      _weights,
-      _trustedWeights,
-      _maxValidatorNumber,
-      _maxPrioritizedValidatorNumber
-    );
+    function _pcPickValidatorSet(
+        address[] memory _candidates,
+        uint256[] memory _weights,
+        uint256[] memory _trustedWeights,
+        uint256 _maxValidatorNumber,
+        uint256 _maxPrioritizedValidatorNumber
+    ) internal pure override returns (address[] memory _result, uint256 _newValidatorCount) {
+        _result = pickValidatorSet(
+            _candidates, _weights, _trustedWeights, _maxValidatorNumber, _maxPrioritizedValidatorNumber
+        );
 
-    _newValidatorCount = _result.length;
-  }
+        _newValidatorCount = _result.length;
+    }
 }
