+++ uniswapv3/deployed-contracts/GovernorBravo/GovernorBravoDelegator.p.sol
@@ -16,7 +16,17 @@ contract GovernorBravoDelegatorStorage {
 
 contract GovernorBravoEvents {
     /// @notice An event emitted when a new proposal is created
-    event ProposalCreated(uint id, address proposer, address[] targets, uint[] values, string[] signatures, bytes[] calldatas, uint startBlock, uint endBlock, string description);
+    event ProposalCreated(
+        uint256 id,
+        address proposer,
+        address[] targets,
+        uint256[] values,
+        string[] signatures,
+        bytes[] calldatas,
+        uint256 startBlock,
+        uint256 endBlock,
+        string description
+    );
 
     /// @notice An event emitted when a vote has been cast on a proposal
     /// @param voter The address which casted a vote
@@ -24,28 +34,28 @@ contract GovernorBravoEvents {
     /// @param support Support value for the vote. 0=against, 1=for, 2=abstain
     /// @param votes Number of votes which were cast by the voter
     /// @param reason The reason given for the vote by the voter
-    event VoteCast(address indexed voter, uint proposalId, uint8 support, uint votes, string reason);
+    event VoteCast(address indexed voter, uint256 proposalId, uint8 support, uint256 votes, string reason);
 
     /// @notice An event emitted when a proposal has been canceled
-    event ProposalCanceled(uint id);
+    event ProposalCanceled(uint256 id);
 
     /// @notice An event emitted when a proposal has been queued in the Timelock
-    event ProposalQueued(uint id, uint eta);
+    event ProposalQueued(uint256 id, uint256 eta);
 
     /// @notice An event emitted when a proposal has been executed in the Timelock
-    event ProposalExecuted(uint id);
+    event ProposalExecuted(uint256 id);
 
     /// @notice An event emitted when the voting delay is set
-    event VotingDelaySet(uint oldVotingDelay, uint newVotingDelay);
+    event VotingDelaySet(uint256 oldVotingDelay, uint256 newVotingDelay);
 
     /// @notice An event emitted when the voting period is set
-    event VotingPeriodSet(uint oldVotingPeriod, uint newVotingPeriod);
+    event VotingPeriodSet(uint256 oldVotingPeriod, uint256 newVotingPeriod);
 
     /// @notice Emitted when implementation is changed
     event NewImplementation(address oldImplementation, address newImplementation);
 
     /// @notice Emitted when proposal threshold is set
-    event ProposalThresholdSet(uint oldProposalThreshold, uint newProposalThreshold);
+    event ProposalThresholdSet(uint256 oldProposalThreshold, uint256 newProposalThreshold);
 
     /// @notice Emitted when pendingAdmin is changed
     event NewPendingAdmin(address oldPendingAdmin, address newPendingAdmin);
@@ -55,38 +65,44 @@ contract GovernorBravoEvents {
 }
 
 contract GovernorBravoDelegator is GovernorBravoDelegatorStorage, GovernorBravoEvents {
-	constructor(
-			address timelock_,
-			address uni_,
-			address admin_,
-	        address implementation_,
-	        uint votingPeriod_,
-	        uint votingDelay_,
-            uint proposalThreshold_) public {
-
+    constructor(
+        address timelock_,
+        address uni_,
+        address admin_,
+        address implementation_,
+        uint256 votingPeriod_,
+        uint256 votingDelay_,
+        uint256 proposalThreshold_
+    ) public {
         // Admin set to msg.sender for initialization
         admin = msg.sender;
 
-        delegateTo(implementation_, abi.encodeWithSignature("initialize(address,address,uint256,uint256,uint256)",
-                                                            timelock_,
-                                                            uni_,
-                                                            votingPeriod_,
-                                                            votingDelay_,
-                                                            proposalThreshold_));
+        delegateTo(
+            implementation_,
+            abi.encodeWithSignature(
+                "initialize(address,address,uint256,uint256,uint256)",
+                timelock_,
+                uni_,
+                votingPeriod_,
+                votingDelay_,
+                proposalThreshold_
+            )
+        );
 
         _setImplementation(implementation_);
 
-		admin = admin_;
-	}
+        admin = admin_;
+    }
 
-
-	/**
+    /**
      * @notice Called by the admin to update the implementation of the delegator
      * @param implementation_ The address of the new implementation for delegation
      */
     function _setImplementation(address implementation_) public {
         require(msg.sender == admin, "GovernorBravoDelegator::_setImplementation: admin only");
-        require(implementation_ != address(0), "GovernorBravoDelegator::_setImplementation: invalid implementation address");
+        require(
+            implementation_ != address(0), "GovernorBravoDelegator::_setImplementation: invalid implementation address"
+        );
 
         address oldImplementation = implementation;
         implementation = implementation_;
@@ -103,28 +119,26 @@ contract GovernorBravoDelegator is GovernorBravoDelegatorStorage, GovernorBravoE
     function delegateTo(address callee, bytes memory data) internal {
         (bool success, bytes memory returnData) = callee.delegatecall(data);
         assembly {
-            if eq(success, 0) {
-                revert(add(returnData, 0x20), returndatasize)
-            }
+            if eq(success, 0) { revert(add(returnData, 0x20), returndatasize) }
         }
     }
 
-	/**
+    /**
      * @dev Delegates execution to an implementation contract.
      * It returns to the external caller whatever the implementation returns
      * or forwards reverts.
      */
-    function () external payable {
+    function() external payable {
         // delegate all other functions to current implementation
-        (bool success, ) = implementation.delegatecall(msg.data);
+        (bool success,) = implementation.delegatecall(msg.data);
 
         assembly {
-              let free_mem_ptr := mload(0x40)
-              returndatacopy(free_mem_ptr, 0, returndatasize)
+            let free_mem_ptr := mload(0x40)
+            returndatacopy(free_mem_ptr, 0, returndatasize)
 
-              switch success
-              case 0 { revert(free_mem_ptr, returndatasize) }
-              default { return(free_mem_ptr, returndatasize) }
+            switch success
+            case 0 { revert(free_mem_ptr, returndatasize) }
+            default { return(free_mem_ptr, returndatasize) }
         }
     }
 }
